split JavaHelper into individual classes
