+++ bnf-language/src/org/intellij/grammar/java/MyElement.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2011-2026 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+ */
+
+package org.intellij.grammar.java;
+
+import com.intellij.psi.NavigatablePsiElement;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.impl.FakePsiElement;
+
+class MyElement<T> extends FakePsiElement implements NavigatablePsiElement {
+
+  final T delegate;
+
+  MyElement(T delegate) {
+    this.delegate = delegate;
+  }
+
+  @Override
+  public PsiElement getParent() {
+    return null;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (o == null || getClass() != o.getClass()) return false;
+
+    MyElement<?> element = (MyElement<?>)o;
+
+    return delegate.equals(element.delegate);
+  }
+
+  @Override
+  public int hashCode() {
+    return delegate.hashCode();
+  }
+
+  @Override
+  public String toString() {
+    return delegate.toString();
+  }
+}
